requirements.txt
# Core Data Science Libraries
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0

# Machine Learning Models
xgboost==1.7.6

# Data Fetching & Exchange APIs
ccxt==4.0.101

# Technical Indicators
ta==0.11.0

# API Framework
fastapi==0.103.1
uvicorn[standard]==0.23.2
pydantic==2.3.0

# Configuration Management
pyyaml==6.0.1

# Logging
loguru==0.7.2

# Utilities
python-dotenv==1.0.0
requests==2.31.0
tqdm==4.66.1
joblib==1.3.2

# Job Scheduling
schedule==1.2.0

# Data Visualization (Optional)
matplotlib==3.7.2
seaborn==0.12.2

# Testing
pytest==7.4.2
httpx==0.24.1

# Development Tools
black==23.9.1
flake8==6.1.0
